llm_assistant/requirements.txt
flask
langchain
langchain-community
langchain-core
langchain-groq
python-dotenv
flask-cors